app/requirements.txt
Flask==1.0.2
PyYAML==5.1
requests==2.21.0

